refactor: consolidate shared types into @odh-dashboard/model-registry/shared
Move DeployPrefillData, ModelDeployPrefillInfo, and RegisteredModelRef
to a canonical shared location in model-registry. Model-serving now
imports these types instead of maintaining local mirrors.

Update Dockerfiles to include model-registry/shared for workspace
resolution in builds that depend on model-serving.
