+++ packages/eval-hub/Dockerfile.workspace
@@ -34,6 +34,8 @@ COPY --chown=default:root packages/foundation/ ./packages/foundation/
 COPY --chown=default:root packages/k8s-core/ ./packages/k8s-core/
 COPY --chown=default:root packages/ui-core/ ./packages/ui-core/
 COPY --chown=default:root packages/model-serving/ ./packages/model-serving/
+COPY --chown=default:root packages/model-registry/package.json ./packages/model-registry/
+COPY --chown=default:root packages/model-registry/shared/ ./packages/model-registry/shared/
 COPY --chown=default:root packages/hardware-profiles/ ./packages/hardware-profiles/
 COPY --chown=default:root packages/connection-types/ ./packages/connection-types/
 # Copy only the frontend source (not the entire frontend build artifacts) for cross-module imports
