+++ packages/model-registry/package.json
@@ -54,6 +54,9 @@
   ],
   "exports": {
     "./extensions": "./extensions.ts",
+    "./shared": {
+      "types": "./shared/index.ts"
+    },
     "./model-catalog-deploy": {
       "types": "./upstream/frontend/src/odh/extension-points/model-catalog-deploy.ts"
     },
