+++ packages/model-registry/upstream/frontend/src/odh/extension-points/model-catalog-deploy.ts
@@ -1,15 +1,8 @@
 import type { Extension, CodeRef } from '@openshift/dynamic-plugin-sdk';
 import { createExtensionGuard } from '@odh-dashboard/plugin-core/extension-points';
+import type { DeployPrefillData } from '@odh-dashboard/model-registry/shared';
 
-export type DeployPrefillData = {
-  modelName: string;
-  modelUri?: string;
-  returnRouteValue?: string;
-  cancelReturnRouteValue?: string;
-  wizardStartIndex?: number;
-  modelType?: 'predictive' | 'generative';
-  prefillAlertText?: string;
-};
+export type { DeployPrefillData } from '@odh-dashboard/model-registry/shared';
 
 export type NavigateToDeploymentWizardWithDataExtension = Extension<
   'model-catalog.deployment/navigate-wizard',
