+++ packages/model-serving/package.json
@@ -72,6 +72,7 @@
     "@odh-dashboard/hardware-profiles": "*",
     "@odh-dashboard/internal": "*",
     "@odh-dashboard/k8s-core": "*",
+    "@odh-dashboard/model-registry": "*",
     "@odh-dashboard/plugin-core": "*",
     "@odh-dashboard/ui-core": "*",
     "lodash-es": "^4.18.1",
